+++ src/main/java/guru/sfg/brewery/repositories/security/UserRepository.java
@@ -2,7 +2,6 @@ package guru.sfg.brewery.repositories.security;
 
 import guru.sfg.brewery.domain.security.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
@@ -12,7 +11,6 @@ import java.util.Optional;
  * Date: Sep 22, 2022
  * Description:
  ***************************/
-public interface UserRepository extends JpaRepository<User, Long> {
-
+public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findByUsername(String username);
 }
